Remove the use of intent provider for request vpn permission
